Changes: Applied to seperate the business logic annd presentation data within home-page
