+++ apps/backoffice-v2/src/pages/Home/Home.page.tsx
@@ -1,31 +1,27 @@
-import React, { FunctionComponent, useEffect } from 'react';
-import { NavLink, Outlet, useLocation, useNavigate } from 'react-router-dom';
-import { t } from 'i18next';
-import { useAuthenticatedUserQuery } from '@/domains/auth/hooks/queries/useAuthenticatedUserQuery/useAuthenticatedUserQuery';
+import React, { FunctionComponent } from 'react';
+import { NavLink, Outlet } from 'react-router-dom';
 import { UserAvatar } from '@/common/components/atoms/UserAvatar/UserAvatar';
 import { Tabs } from '@/common/components/organisms/Tabs/Tabs';
 import { TabsList } from '@/common/components/organisms/Tabs/Tabs.List';
 import { TabsTrigger } from '@/common/components/organisms/Tabs/Tabs.Trigger';
 import { TabsContent } from '@/common/components/organisms/Tabs/Tabs.Content';
-import { useLocale } from '@/common/hooks/useLocale/useLocale';
 import { DateRangePicker } from '@/common/components/molecules/DateRangePicker/DateRangePicker';
 import { useHomeLogic } from '@/common/hooks/useHomeLogic/useHomeLogic';
 
 export const Home: FunctionComponent = () => {
-  const { data: session } = useAuthenticatedUserQuery();
-  const { firstName, fullName, avatarUrl } = session?.user || {};
-  const locale = useLocale();
-  const { pathname } = useLocation();
-  const navigate = useNavigate();
-  const { handleDateRangeChange, from, to } = useHomeLogic();
-
-  useEffect(() => {
-    if (pathname !== `/${locale}` && pathname !== `/${locale}/home`) {
-      return;
-    }
-
-    navigate(`/${locale}/home/statistics`);
-  }, [pathname, locale, navigate]);
+  const {
+    handleDateRangeChange,
+    from,
+    to,
+    t,
+    firstName,
+    fullName,
+    avatarUrl,
+    statisticsLink,
+    workflowsLink,
+    pathname,
+    locale,
+  } = useHomeLogic();
 
   return (
     <div className={`flex flex-col gap-10 p-10`}>
@@ -50,10 +46,10 @@ export const Home: FunctionComponent = () => {
         <Tabs defaultValue={pathname} key={pathname}>
           <TabsList>
             <TabsTrigger asChild={true} value={`/${locale}/home/statistics`}>
-              <NavLink to={`/${locale}/home/statistics`}>Statistics</NavLink>
+              <NavLink to={statisticsLink}>Statistics</NavLink>
             </TabsTrigger>
             <TabsTrigger asChild={true} value={`/${locale}/home/workflows`}>
-              <NavLink to={`/${locale}/home/workflows`}>Workflows</NavLink>
+              <NavLink to={workflowsLink}>Workflows</NavLink>
             </TabsTrigger>
           </TabsList>
           <TabsContent value={pathname}>
